Created Gameplay Loop
I have abstarcted the gameplay loop into a new class which will allow for the inclusion of a try catch to allow for errors to be detected and managed properly.

Next I will need to create a basic game board for the card game; probably by using simple squares/rectangles.