Added mouse control into the game.
I have added the mouse class into the game. I have used the same class from my final year project but I have also added the functionality of a mouse cureser which will follow the mouse around the window.

Next is to add an opponent into the game which will at the moment hold the position of their mouse curser.