Allowed for basic networking
I have allowed for two open solutions to connect upon startup, however this seems to close the second solution. I will need to find a work around for that but for now it looks to be working.

Next, I plan to pass the location data for the mouse onto to the other player to cement my understanding of sfml networking using packets.