Added more cards into collection
I have added more cards into the collection which are all loaded properly, as well as passing the data into the card objects.

I have also moved the prining of the card details from withing the loading of the collection into the cards themselves allowing for the checking of the data within them easily.

Next, I aim to draw the cards with all of their data into the game window.