Resource Manager
I have added a basic template class which I will use to manage the textures and the fonts within the game.

the main focus of the map will be the variable type the template is initalized with, so if you want to manahe fonts intilize the tempate with an sf::font.