Box with text inside
I have created a class which will allow for a box to be created and text will be placed within it, I have used the width for the text rather than the height so if there is a large paragraph then the sizing may need to be adjusted slightly.

But this allows for me to place a box where I need it and the text will fit perfectly within it.