Cards and card abilities
I have placed all of the card info onto the cards. This required me to add a text wrapping function which I added in addition to the text sizing part of the card generation. This will allow for the card information to be split over multiple lines if nessecary. This will allow for the text sizes to be larger, thus easier to read with more information being displayed.

Next, I will need to work on the main gameplay loop, the cards on the board interacting with each other.