Begun work on card targeting
I have laid the groundworks for the card targeting in the form of a combat loop class which holds a combat setup step which should house the targeting section of the game.

I have added more information into the XML file in the form of two additional nodes for each ability (so four in total) which will hold the information on how the ability will work (see comments on xml for more info), and the way the ability will target cards.

Next, I will begin with the card targeting with the basic sigle targets with the expectation to add other targeting methods at a later point.